requirements.txt
django
django-debug-toolbar
django-model-utils
psycopg2